finish main rebase adjustments
